[Refactor] WorkDetail의 로직과 뷰를 분리
점점 비대해지는 WorkDetail 을 유지보수하기 힘들 것 같아서
WorksDetailContainer를 만듬 -> 로직
WorkDetail -> 뷰만 담당

추후 WorkDetail을 분리하여 리렌더링되는 범위또한 조절 가능
