+++ client/src/components/FeedContainer/index.tsx
@@ -7,13 +7,14 @@ import UploadMain from '../UploadMain';
 import FeedWallpapers from '../FeedWallpapers';
 import FeedMusic from '../FeedMusics';
 import FeedWorks from '../FeedWorks';
-import WorkDetail from '../WorkDetail';
 import FeedNavigator from '../FeedNavigator';
 import NotFound from '../../components/NotFound';
 
+import WorkDetailContainer from '../../containers/WorkDetailContainer';
+
 const Content = ({ match }: RouteComponentProps) => (
   <Switch>
-    <Route path={`${match.path}/detail-image/:id`} component={WorkDetail} />
+    <Route path={`${match.path}/detail-image/:id`} component={WorkDetailContainer} />
     <Route path={`${match.path}`}>
       <Route path={`${match.path}`} component={FeedNavigator} />
       <Switch>
