+++ client/src/components/WorksDetail/styles.ts
