+++ client/src/hooks/useGetFeedList.ts
@@ -25,6 +25,7 @@ const useFetch = <T>(initData:T[])
           setIsError(false);
         }
       } catch (e) {
+        console.log(e);
         setIsError(true);
       }
     };
